Refactoring project Folders
