+++ src/components/App.js
@@ -1,19 +1,22 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
-import SignUp from "./pages/Sign-up";
+import SignUp from "../pages/Sign-up";
 import GlobalStyle from "../styles/reset";
-import Timeline from "./pages/Timeline.js";
-import SignIn from "./pages/Sign-in";
-import UserTimeline from "./pages/UserTimeline";
-import Hashtag from "./pages/Hashtag";
+import Timeline from "../pages/Timeline";
+import SignIn from "../pages/Sign-in";
+import UserTimeline from "../pages/UserPosts";
+import Hashtag from "../pages/Hashtag";
 import LoginContext from "../contexts/LoginContext";
 import { useState } from "react";
-import PrivatePage from "./commons/PrivatePage";
-import { getUser } from "../services/localstorage";
+import { getUser } from "../utils/localstorage";
+import PrivatePage from "../decorator/PrivatePage";
 
 export default function App() {
   const [userData, setUserData] = useState(getUser());
   const [config, setConfig] = useState(null);
+  const PrivateTimeLine = <PrivatePage> <Timeline/> </PrivatePage>
+  const PrivateHashtag =  <PrivatePage> <Hashtag /> </PrivatePage>
+  const PrivateUserPosts = <PrivatePage>  <UserTimeline /> </PrivatePage>
 
   return (
     <BrowserRouter>
@@ -29,21 +32,9 @@ export default function App() {
         <Routes>
           <Route path="/" element={<SignIn />} />
           <Route path="/sign-up" element={<SignUp />} />
-          <Route path="/timeline" element={
-            <PrivatePage>
-                <Timeline />
-            </PrivatePage>
-          } />
-          <Route path="/hashtag/:hashtag" element={
-            <PrivatePage>
-               <Hashtag />
-          </PrivatePage>
-          } />
-          <Route path="/user/:id" element={
-            <PrivatePage>
-               <UserTimeline />
-            </PrivatePage>
-          } />
+          <Route path="/timeline" element={PrivateTimeLine} />
+          <Route path="/hashtag/:hashtag" element={PrivateHashtag} />
+          <Route path="/user/:id" element={PrivateUserPosts} />
         </Routes>
       </LoginContext.Provider>
     </BrowserRouter>
