+++ src/components/commons/Navbar.js
@@ -21,11 +21,12 @@ export default function Navbar() {
 
     getUserByName({config, search})
       .then((res) => {
-        console.log(res.data);
         setUsersSearch(res.data);
       })
       .catch((res) => {
-        console.log(res);
+        if(res.response.status === 422) {
+          setHiddenSearch(true);
+        };
       })
   }, [search, config])
