+++ src/components/post/Icon-like.js
@@ -76,7 +76,7 @@ export default function IconLike(postId) {
 
 	return (
 		<LikesStyle>
-			{isLiked ? (
+			{isLiked === true ? (
 				<AiFillHeart color="red" onClick={() => likeClick()} />
 			) : (
 				<TbHeart onClick={() => likeClick()} />
