+++ src/pages/Sign-up.js
@@ -1,11 +1,13 @@
 import { useState } from "react";
+import { ThreeDots } from "react-loader-spinner";
 
 import { useNavigate } from "react-router-dom";
+import FormPage from "../components/initialPages/Form-page";
+import { signUp } from "../services/axios";
+import FormStyle from "../styles/commons/Form";
 
-import FormPage from "../Form-page";
-import FormStyle from "../../styles/commons/Form";
-import { signUp } from "../../services/axios";
-import { ThreeDots } from "react-loader-spinner";
+
+
 
 function Inputs({ handleForm, form, disabled }) {
   return (
@@ -73,8 +75,12 @@ export default function SignUp() {
       .catch((answer) => {
         setDisabled(false);
         if (answer.response.status === 409) {
-          alert("Email já em uso");
+          alert("E-mail already in use");
         }
+        if (answer.response.status === 422) {
+          alert("Invalid Format");
+        }
+
       });
   }
   function handleForm(e) {
@@ -84,6 +90,7 @@ export default function SignUp() {
     });
   }
   return (
+  
     <FormPage>
       <FormStyle onSubmit={submitData}>
         <Inputs handleForm={handleForm} disabled={disabled} form={form} />
