+++ src/styles/posts/DeletePost.js
