+++ src/styles/publishbox/PublishContent.js
